+++ rails_root/app/controllers/survey_profiles_controller.rb
@@ -70,8 +70,13 @@ class SurveyProfilesController < ApplicationController
     @survey_profile.destroy!
 
     respond_to do |format|
-      format.html { redirect_to survey_profiles_url, notice: 'Survey profile was successfully destroyed.' }
-      format.json { head :no_content }
+      if user_is_admin?
+        format.html { redirect_to admin_dashboard_path, notice: 'Survey profile was successfully destroyed.' }
+        format.json { head :no_content }
+      else
+        format.html { redirect_to root_path, notice: 'Survey profile was successfully destroyed.' }
+        format.json { head :no_content }
+      end
     end
   end
 
@@ -122,20 +127,20 @@ class SurveyProfilesController < ApplicationController
 
   def claim_invitation
     temporary_invitation_session_var = session[:invitation]
-
-    if temporary_invitation_session_var && temporary_invitation_session_var['expiration'] > Time.now
+  
+    if temporary_invitation_session_var
       invitation = Invitation.find_by(id: temporary_invitation_session_var['from'])
+      
       if invitation
         sharecode_from_invitation = invitation.parent_response.share_code
-        # survey_profile = SurveyProfile.find_by(user_id: session[:userinfo]['sub'])
         new_response_to_fill = SurveyResponse.create(profile: @survey_profile, share_code: sharecode_from_invitation)
         invitation.update(claimed_by_id: @survey_profile.id, response_id: new_response_to_fill.id)
         redirect_to edit_survey_response_path(new_response_to_fill)
         return true
       end
     end
-
-    session.delete(:invitation)
+  
+    session.delete(:invitation) if temporary_invitation_session_var.nil?
     false
-  end
+  end  
 end
